python/image/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
pandas==2.0.3
Pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3
scikit-learn==1.3.0
PaddlePaddle==2.5.2
PaddleOCR==2.7.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
requests==2.31.0
firebase-admin==6.2.0
pyjwt==2.8.0
cryptography==41.0.4